+++ src/solver/optimisation/opt_appel_solveur_lineaire.cpp
@@ -20,7 +20,6 @@
  */
 
 #include <chrono>
-#include <span>
 
 #include <antares/antares/fatal-error.h>
 #include <antares/logs/logs.h>