+++ docs/developer-guide/CHANGELOG.md
@@ -152,6 +152,16 @@ toc_depth: 2
 
 ## Branch 9.2.x
 
+### 9.2.3
+#### Bugfixes
+* Fix segfault when exporting binding constraints results (#2954)
+
+### 9.2.2
+#### New features
+> [!CAUTION] Breaking change
+
+* [ANT-3666] Move STS additional constraint data folder (#2945)
+
 ### 9.2.1
 #### New features
 * Add new near price cap hours variable (NPCAP HOURS) [ANT-3038] (#2815)
@@ -490,6 +500,10 @@ toc_depth: 2
 * Fix invalid index causing segfault in `test-study` test (#1902)
 
 ## Branch 8.8.x (end of support 12/2025)
+
+### 8.8.18 (08/2025)
+#### Bugfixes
+* Fix segfault when exporting binding constraints results (#2955)
 
 ### 8.8.17 (06/2025)
 #### Bugfixes